:truck: Move app into components
